Move CSS into local style files
Should keep it more contained and ensure old CSS doesn’t hang around
