+++ client/component/modal/style.scss
@@ -1,3 +1,5 @@
+@import '../../globals';
+
 body.redirection-modal {
 	overflow: hidden;
 }
