+++ client/component/wordpress/placeholder.scss
@@ -1,8 +1,10 @@
+@import '../../globals';
+
 /**
  * Inline loading styles
  */
 
-@keyframes loading-fade {
+ @keyframes loading-fade {
 	0% { opacity: .5; }
 	50% { opacity: 1; }
 	100% { opacity: .5; }
