+++ client/component/wordpress/spinner.scss
