+++ client/page/io/style.scss
@@ -1,7 +1,7 @@
 /**
  * Dropzone area on import/export page
  */
-.dropzone {
+ .dropzone {
 	border: 3px dashed #bbb;
 	text-align: center;
 	padding: 10px;
