+++ client/page/logs/delete-all.js
