+++ client/page/logs/index.js
@@ -14,8 +14,8 @@ import Table from 'component/table';
 import TableNav from 'component/table/navigation';
 import SearchBox from 'component/table/search';
 import { LOGS_TYPE_REDIRECT } from 'state/log/type';
-import DeleteAll from 'component/logs/delete-all';
-import ExportCSV from 'component/logs/export-csv';
+import DeleteAll from 'page/logs/delete-all';
+import ExportCSV from 'page/logs/export-csv';
 import { tableKey } from 'lib/table';
 import LogRow from './row';
 import { STATUS_COMPLETE, STATUS_IN_PROGRESS, STATUS_SAVING } from 'state/settings/type';
