+++ client/page/logs/row.js
