+++ client/page/logs404/index.js
@@ -15,8 +15,8 @@ import Table from 'component/table';
 import TableNav from 'component/table/navigation';
 import SearchBox from 'component/table/search';
 import TableGroup from 'component/table/group';
-import DeleteAll from 'component/logs/delete-all';
-import ExportCSV from 'component/logs/export-csv';
+import DeleteAll from 'page/logs/delete-all';
+import ExportCSV from 'page/logs/export-csv';
 import Row404 from './row';
 import RowUrl from './row-url';
 import RowIp from './row-ip';
