+++ client/page/options/delete-plugin.js
