+++ client/page/options/donation.js
@@ -10,6 +10,7 @@ import PropTypes from 'prop-types';
  * Internal dependencies
  */
 import { FormTable, TableRow } from 'component/wordpress/form-table';
+import './donation.scss';
 
 const MIN = 16;
 const MAX = 100;
