+++ client/page/options/donation.scss
@@ -1,7 +1,7 @@
 /**
  * The donation section of the options page
  */
-.donation {
+ .donation {
 	.donation-amount {
 		float: left;
 		margin-top: 10px;
