+++ client/page/options/index.js
@@ -11,7 +11,7 @@ import { connect } from 'react-redux';
 import { loadSettings, deletePlugin } from 'state/settings/action';
 import { STATUS_IN_PROGRESS, STATUS_COMPLETE } from 'state/settings/type';
 import OptionsForm from './options-form';
-import DeletePlugin from 'component/options/delete-plugin';
+import DeletePlugin from 'page/options/delete-plugin';
 import Placeholder from 'component/wordpress/placeholder';
 import Donation from './donation';
