+++ client/page/support/status.scss
@@ -1,7 +1,9 @@
+@import '../../globals';
+
 /**
  * The plugin status section of the support page
  */
-.plugin-status {
+ .plugin-status {
 	th {
 		text-align: left;
 		padding: 5px;
